Change VUV majorValue for better compatibility
I propose changing the `majorValue` of the VUV currency to be in consistent format with all other values. That is, to use the value 1.0 to indicate that the minor unit does not have a different factor from the major unit. Previously, this was represented by an empty string "", which could cause compatibility issues with some code.